rename control_pump to controllable_pump
